Disentangle pk_keys.h and pubkey.h with fwd declares
