+++ src/lib/x509/x509_obj.cpp
@@ -10,7 +10,7 @@
 #include <botan/ber_dec.h>
 #include <botan/der_enc.h>
 #include <botan/pem.h>
-#include <botan/pubkey.h>
+#include <botan/pk_keys.h>
 #include <botan/internal/emsa.h>
 #include <botan/internal/fmt.h>
 #include <algorithm>
